Lab3 uploaded
Note: Task 1,2,3 are combined in a folder. And task 4 is in its separate folder. For task 1-3 all the PHP files are connected starting from login.php
After the login a "New Page" button appears and it will take to a new page. There are two buttons to navigate in this page to move to "Changing password" and "Uploading Image"